아이템과 아이템 효과, 플레이어 습득 추가
Former-commit-id: 796c7924ab127094c8a1726052c594d83e33afad

+++ Assets/1.Scripts/Weapon/Gun.cs
@@ -28,7 +28,7 @@ public class Gun : Weapon
                 IDamagable damagable = hit.collider.GetComponent<IDamagable>();
                 //attacker의 스탯과 무기 damage를 계산
                 float totalDamage = CombatSystem.CalculateInflictedDamage(args.Attacker, this, damagable);
-                if (damagable != null) damagable.TakeHit(totalDamage);
+                if (damagable != null) damagable.TakeHit((int)totalDamage);
             }
             _remainShootTerm = ShootTerm;
